requirements.txt
streamlit
crewai
langchain
langchain-google-genai
python-dotenv
python-docx
python-pptx
sqlite3
pydantic
